pyproject.toml
[tool.poetry]
name = "assignment-2"
version = "0.1.0"
description = "assignment2"
authors = ["vishodhan-krishnan <krishnan.vis@northeastern.edu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
sqlalchemy = ">=1.4.36,<2.0"
fastapi = "^0.95.0"
bcrypt = "^4.2.0"
openai = "^1.51.0"
streamlit = "^1.39.0"
mysql-connector-python = "^9.0.0"
pandas = "^2.2.3"
pymupdf = "^1.24.11"
env = "^0.1.0"
uvicorn = "^0.23.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
docker = "^7.1.0"
google-cloud-documentai = "^2.32.0"
tabulate = "^0.9.0"
google-cloud = "^0.34.0"
google-api-python-client = "^2.147.0"
boto3 = "^1.35.36"
apache-airflow = "2.10.2"
python-jose = "^3.3.0"
python-dotenv = "^1.0.0"
snowflake-connector-python = "^3.12.2"
selenium = "^4.25.0"
webdriver-manager = "^4.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"